+++ Cargo.toml
@@ -60,9 +60,9 @@ rome_migrate                = { path = "./crates/rome_migrate" }
 rome_parser                 = { version = "0.0.1", path = "./crates/rome_parser" }
 rome_rowan                  = { version = "0.0.1", path = "./crates/rome_rowan" }
 rome_service                = { path = "./crates/rome_service" }
+rome_suppression            = { version = "0.0.1", path = "./crates/rome_suppression" }
 rome_text_edit              = { version = "0.0.1", path = "./crates/rome_text_edit" }
 rome_text_size              = { version = "0.0.1", path = "./crates/rome_text_size" }
-rome_suppression              = { version = "0.0.1", path = "./crates/rome_suppression" }
 tests_macros                = { path = "./crates/tests_macros" }
 
 # Crates needed in the workspace
