+++ crates/rome_json_formatter/Cargo.toml
@@ -13,10 +13,10 @@ version              = "0.0.1"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+rome_diagnostics = { workspace = true }
 rome_formatter   = { workspace = true }
 rome_json_syntax = { workspace = true }
 rome_rowan       = { workspace = true }
-rome_diagnostics    = { workspace = true }
 rome_suppression = { workspace = true }
 
 [dev-dependencies]
