+++ crates/rome_json_formatter/src/lib.rs
@@ -11,7 +11,7 @@ use crate::comments::JsonCommentStyle;
 pub(crate) use crate::context::JsonFormatContext;
 use crate::context::JsonFormatOptions;
 use crate::cst::FormatJsonSyntaxNode;
-use rome_formatter::comments::{CommentKind, CommentStyle, Comments};
+use rome_formatter::comments::Comments;
 use rome_formatter::prelude::*;
 use rome_formatter::{
     write, CstFormatContext, FormatContext, FormatLanguage, FormatOwnedWithRule, FormatRefWithRule,
@@ -19,7 +19,7 @@ use rome_formatter::{
 };
 use rome_formatter::{Formatted, Printed};
 use rome_json_syntax::{AnyJsonValue, JsonLanguage, JsonSyntaxNode, JsonSyntaxToken};
-use rome_rowan::{AstNode, SyntaxNode, SyntaxTriviaPieceComments, TextRange};
+use rome_rowan::{AstNode, SyntaxNode, TextRange};
 
 /// Used to get an object that knows how to format this object.
 pub(crate) trait AsFormat<Context> {
