+++ crates/rome_suppression/Cargo.toml
@@ -1,14 +1,14 @@
 [package]
-authors= { workspace = true }
-description          = "Rome suppression utilities"
-edition= { workspace = true }
-license              = { workspace = true }
-name                 = "rome_suppression"
-repository= { workspace = true }
-version              = "0.0.1"
+authors     = { workspace = true }
+description = "Rome suppression utilities"
+edition     = { workspace = true }
+license     = { workspace = true }
+name        = "rome_suppression"
+repository  = { workspace = true }
+version     = "0.0.1"
 
 
 [dependencies]
-rome_rowan =  { workspace = true }
-rome_diagnostics =  { workspace = true }
-rome_console =  { workspace = true }
+rome_console     = { workspace = true }
+rome_diagnostics = { workspace = true }
+rome_rowan       = { workspace = true }
